+++ curve25519-dalek/CHANGELOG.md
@@ -5,6 +5,10 @@ major series.
 
 ## 4.x series
 
+### Unreleased
+
+* Add `EdwardsPoint::vartime_check_double_scalar_mul_basepoint`.
+
 ### 4.1.2
 
 * Fix nightly SIMD build